requirements.txt
streamlit
streamlit-webrtc
av
ultralytics
numpy
opencv-python-headless
Pillow